requirements.txt
django==4.2.25
djangorestframework==3.16.1
django-fsm==3.0.1
django-fsm-log==3.1.0
pytest==8.4.2
pytest-django==4.11.1
pytest-mock==3.15.1
